[WIP] Move photo handling from PhotoBuilder to AndroidAddressBook and implement deleteGroupsWithoutMembers
