+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/mapping/contacts/builder/PhotoBuilderTest.kt
@@ -5,25 +5,10 @@
 package at.bitfire.synctools.mapping.contacts.builder
 
 import android.Manifest
-import android.content.ContentProviderClient
-import android.content.ContentUris
-import android.graphics.BitmapFactory
 import android.net.Uri
-import android.provider.ContactsContract
-import android.provider.ContactsContract.RawContacts
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.rule.GrantPermissionRule
 import at.bitfire.synctools.mapping.contacts.Contact
-import at.bitfire.synctools.mapping.contacts.TestUtils
-import at.bitfire.synctools.storage.contacts.AndroidAddressBook
-import at.bitfire.synctools.storage.contacts.AndroidContact
-import at.bitfire.synctools.storage.contacts.TestAddressBook
-import org.junit.AfterClass
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNotNull
-import org.junit.Assert.assertNull
-import org.junit.Assert.assertTrue
-import org.junit.BeforeClass
 import org.junit.ClassRule
 import org.junit.Test
 import kotlin.random.Random
@@ -34,27 +19,6 @@ class PhotoBuilderTest {
         @JvmField
         @ClassRule
         val permissionRule = GrantPermissionRule.grant(Manifest.permission.READ_CONTACTS, Manifest.permission.WRITE_CONTACTS)!!
-
-        val testContext = InstrumentationRegistry.getInstrumentation().context
-        private lateinit var provider: ContentProviderClient
-        private lateinit var addressBook: AndroidAddressBook
-
-        @BeforeClass
-        @JvmStatic
-        fun connect() {
-            provider = testContext.contentResolver.acquireContentProviderClient(ContactsContract.AUTHORITY)!!
-            assertNotNull(provider)
-
-            addressBook = TestAddressBook.create(provider)
-        }
-
-        @AfterClass
-        @JvmStatic
-        fun disconnect() {
-            TestAddressBook.remove(addressBook)
-            @Suppress("DEPRECATION")
-            provider.release()
-        }
     }
 
 
@@ -71,55 +35,8 @@ class PhotoBuilderTest {
         PhotoBuilder(Uri.EMPTY, null, Contact().apply {
             photo = blob
         }, false).build().also { result ->
-            // no row because photos have to be inserted with a separate call to insertPhoto()
+            // no row because photos have to be set with a separate call to AndroidAddressBook.setPhoto()
             assertEquals(0, result.size)
-        }
-    }
-
-
-    @Test
-    fun testInsertPhoto() {
-        val contact = AndroidContact(addressBook, Contact().apply { displayName = "Contact with photo" }, null, null)
-        val contactUri = contact.add()
-        val rawContactId = ContentUris.parseId(contactUri)
-
-        try {
-            val photo = TestUtils.resourceToByteArray("/large.jpg")
-            val photoUri = PhotoBuilder.insertPhoto(provider, rawContactId, photo)
-            assertNotNull(photoUri)
-
-            // the photo is processed and often resized by the contacts provider
-            val contact2 = addressBook.findContactById(rawContactId)
-            val photo2 = contact2.getContact().photo!!
-
-            // verify that the image is in JPEG format (some Samsung devices seem to save as PNG)
-            val options = BitmapFactory.Options()
-            options.inJustDecodeBounds = true
-            BitmapFactory.decodeByteArray(photo2, 0, photo2.size, options)
-            assertEquals("image/jpeg", options.outMimeType)
-
-            // verify that contact is not dirty
-            provider.query(
-                ContentUris.withAppendedId(RawContacts.CONTENT_URI, rawContactId),
-                arrayOf(RawContacts.DIRTY),
-                null, null, null
-            )!!.use { cursor ->
-                assertTrue(cursor.moveToNext())
-                assertEquals(0, cursor.getInt(0))
-            }
-        } finally {
-            contact.delete()
-        }
-    }
-
-    @Test
-    fun testInsertPhoto_Invalid() {
-        val contact = AndroidContact(addressBook, Contact().apply { displayName = "Contact with photo" }, null, null)
-        contact.add()
-        try {
-            assertNull(PhotoBuilder.insertPhoto(provider, contact.id!!, ByteArray(100) /* invalid photo  */))
-        } finally {
-            contact.delete()
         }
     }
